recommendation-system/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.3.0
redis==5.0.0
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.1.0
scipy==1.11.0
joblib==1.3.0